+++ packages/orq-rc/src/models/components/reasoningpart.ts
@@ -40,7 +40,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kc6jynt9ts4szb85drearx3j"),
+  _id: z.string().default("reasoning_01kc8fgre1adjf755q2cfn9p5b"),
   metadata: z.record(z.any()).optional(),
   kind: z.literal("reasoning"),
   reasoning: z.string(),
@@ -65,7 +65,7 @@ export const ReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kc6jynt9ts4szb85drearx3j"),
+  id: z.string().default("reasoning_01kc8fgre1adjf755q2cfn9p5b"),
   metadata: z.record(z.any()).optional(),
   kind: z.literal("reasoning"),
   reasoning: z.string(),