+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -196,7 +196,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-12-11T11:32:37.634Z",
+    "2025-12-12T05:11:03.686Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -229,7 +229,7 @@ export const FileUploadResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-12-11T11:32:37.634Z"))
+  created: z.date().default(() => new Date("2025-12-12T05:11:03.686Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {