+++ packages/orq-rc/src/models/operations/getbudget.ts
@@ -295,7 +295,7 @@ export const GetBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-12-11T11:32:34.996Z",
+    "2025-12-12T05:10:58.477Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -335,7 +335,7 @@ export const GetBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => GetBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-12-11T11:32:34.996Z"))
+  updated: z.date().default(() => new Date("2025-12-12T05:10:58.477Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {