+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3973,7 +3973,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KC6JYP0T71EE54ZH9RVVZWMV"),
+  id: z.string().default("01KC8FGRMYM3VJ9EZZQGSV5TXA"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -3992,7 +3992,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KC6JYP0T71EE54ZH9RVVZWMV"),
+  id: z.string().default("01KC8FGRMYM3VJ9EZZQGSV5TXA"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),