+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -335,7 +335,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kc6jyp2nnktr6hwd6nxvx3ab"),
+  _id: z.string().default("conv_01kc8fgrpzpkv1yesxr9j73gfm"),
   entity_id: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   display_name: z.string(),
@@ -375,7 +375,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kc6jyp2nnktr6hwd6nxvx3ab"),
+  id: z.string().default("conv_01kc8fgrpzpkv1yesxr9j73gfm"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),