+++ packages/orq-rc/RELEASES.md
@@ -5539,3 +5539,13 @@ Based on:
 - [typescript v4.1.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.38 - packages/orq-rc
+
+## 2025-12-12 05:13:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.675.4 (2.779.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.39 - packages/orq-rc